Move arithmetization files
