+++ barretenberg/cpp/src/barretenberg/plonk_honk_shared/execution_trace/execution_trace_block.hpp
