+++ barretenberg/cpp/src/barretenberg/plonk_honk_shared/execution_trace/gate_data.hpp
