+++ barretenberg/cpp/src/barretenberg/stdlib_circuit_builders/standard_circuit_builder.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include "barretenberg/ecc/curves/bn254/bn254.hpp"
 #include "barretenberg/ecc/curves/grumpkin/grumpkin.hpp"
-#include "barretenberg/plonk_honk_shared/arithmetization/standard_arithmetization.hpp"
+#include "barretenberg/plonk_honk_shared/execution_trace/standard_execution_trace.hpp"
 #include "barretenberg/plonk_honk_shared/types/circuit_type.hpp"
 #include "barretenberg/plonk_honk_shared/types/merkle_hash_type.hpp"
 #include "barretenberg/plonk_honk_shared/types/pedersen_commitment_type.hpp"
